update for yolov2 support
