+++ dynamic_vino_lib/include/dynamic_vino_lib/inferences/object_detection.hpp
@@ -16,9 +16,8 @@
  * @brief A header file with declaration for ObjectDetection Class
  * @file object_detection.hpp
  */
-#ifndef DYNAMIC_VINO_LIB_INFERENCES_OBJECT_DETECTION_YOLOV2_HPP_
-#define DYNAMIC_VINO_LIB_INFERENCES_OBJECT_DETECTION_YOLOV2_HPP_
-
+#ifndef DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
+#define DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
 #include <object_msgs/msg/object.hpp>
 #include <object_msgs/msg/object_in_box.hpp>
 #include <object_msgs/msg/objects_in_boxes.hpp>
@@ -34,16 +33,18 @@
 #include "inference_engine.hpp"
 #include "opencv2/opencv.hpp"
 // namespace
-namespace dynamic_vino_lib {
+namespace dynamic_vino_lib
+{
 /**
- * @class ObjectDetectionYolov2voc
+ * @class ObjectDetectionSSD
  * @brief Class to load face detection model and perform face detection.
  */
-class ObjectDetectionYolov2 : public ObjectDetection {
- public:
+class ObjectDetectionSSD : public ObjectDetection
+{
+public:
   using Result = dynamic_vino_lib::ObjectDetectionResult;
-  explicit ObjectDetectionYolov2(double);
-  ~ObjectDetectionYolov2() override;
+  explicit ObjectDetectionSSD(bool, double);
+  ~ObjectDetectionSSD() override;
   /**
    * @brief Load the face detection model.
    */
@@ -56,12 +57,8 @@ class ObjectDetectionYolov2 : public ObjectDetection {
    * to the frame generated by the input device.
    * @return Whether this operation is successful.
    */
-  bool enqueue(const cv::Mat&, const cv::Rect&) override;
-  /**
-   * @brief Start inference for all buffered frames.
-   * @return Whether this operation is successful.
-   */
-  bool submitRequest() override;
+  bool enqueue(const cv::Mat &, const cv::Rect &) override;
+
   /**
    * @brief This function will fetch the results of the previous inference and
    * stores the results in a result buffer array. All buffered frames will be
@@ -79,35 +76,34 @@ class ObjectDetectionYolov2 : public ObjectDetection {
    * to the frame generated by the input device.
    * @param[in] idx The index of the result.
    */
-  const dynamic_vino_lib::Result* getLocationResult(int idx) const override;
+  const dynamic_vino_lib::Result * getLocationResult(int idx) const override;
   /**
    * @brief Show the observed detection result either through image window
      or ROS topic.
    */
-  const void observeOutput(const std::shared_ptr<Outputs::BaseOutput>& output);
+  const void observeOutput(const std::shared_ptr<Outputs::BaseOutput> & output);
   /**
    * @brief Get the name of the Inference instance.
    * @return The name of the Inference instance.
    */
   const std::string getName() const override;
 
- private:
-  std::shared_ptr<Models::ObjectDetectionYolov2Model> valid_model_;
-  std::vector<Result> raw_results_;
+  /**
+   * @brief Calculate the IoU ratio for the given rectangles.
+   * @return IoU Ratio of the given rectangles.
+   */
+  static double IntersectionOverUnion(const cv::Rect &box_1, const cv::Rect &box_2);
+
+private:
+  std::shared_ptr<Models::ObjectDetectionModel> valid_model_;
   std::vector<Result> results_;
-  int width_ = 0;
-  int height_ = 0;
-  int max_proposal_count_;
-  int object_size_;
+
   double show_output_thresh_ = 0;
 
-  bool matToBlob(const cv::Mat& frame, const cv::Rect&, float scale_factor,
-                 int batch_index, const std::string& input_name);
-  int getEntryIndex(int side, int lcoords, int lclasses, int location, int entry);
-  double IntersectionOverUnion(const Result &box_1, const Result &box_2);
-  //int enqueued_frames_ = 0;
   int max_batch_size_ = 1;
-  cv::Rect ROI_;
+
+  bool enable_roi_constraint_ = false;
+
 };
 }  // namespace dynamic_vino_lib
-#endif  // DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_YOLOV2_HPP_
+#endif  // DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
