--- dynamic_vino_lib/include/dynamic_vino_lib/inferences/object_detection_ssd.hpp
@@ -1,116 +1,0 @@
-// Copyright (c) 2018 Intel Corporation
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @brief A header file with declaration for ObjectDetection Class
- * @file object_detection.hpp
- */
-#ifndef DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
-#define DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
-#include <object_msgs/msg/object.hpp>
-#include <object_msgs/msg/object_in_box.hpp>
-#include <object_msgs/msg/objects_in_boxes.hpp>
-#include <object_msgs/srv/detect_object.hpp>
-#include <rclcpp/rclcpp.hpp>
-#include <memory>
-#include <vector>
-#include <string>
-#include "dynamic_vino_lib/models/object_detection_ssd_model.hpp"
-#include "dynamic_vino_lib/models/object_detection_yolov2_model.hpp"
-#include "dynamic_vino_lib/engines/engine.hpp"
-#include "dynamic_vino_lib/inferences/base_inference.hpp"
-#include "inference_engine.hpp"
-#include "opencv2/opencv.hpp"
-// namespace
-namespace dynamic_vino_lib
-{
-/**
- * @class ObjectDetectionSSD
- * @brief Class to load face detection model and perform face detection.
- */
-class ObjectDetectionSSD : public ObjectDetection
-{
-public:
-  using Result = dynamic_vino_lib::ObjectDetectionResult;
-  explicit ObjectDetectionSSD(bool, double);
-  ~ObjectDetectionSSD() override;
-  /**
-   * @brief Load the face detection model.
-   */
-  void loadNetwork(std::shared_ptr<Models::ObjectDetectionModel>) override;
-  /**
-   * @brief Enqueue a frame to this class.
-   * The frame will be buffered but not infered yet.
-   * @param[in] frame The frame to be enqueued.
-   * @param[in] input_frame_loc The location of the enqueued frame with respect
-   * to the frame generated by the input device.
-   * @return Whether this operation is successful.
-   */
-  bool enqueue(const cv::Mat &, const cv::Rect &) override;
-  /**
-   * @brief Start inference for all buffered frames.
-   * @return Whether this operation is successful.
-   */
-  bool submitRequest() override;
-  /**
-   * @brief This function will fetch the results of the previous inference and
-   * stores the results in a result buffer array. All buffered frames will be
-   * cleared.
-   * @return Whether the Inference object fetches a result this time
-   */
-  bool fetchResults() override;
-  /**
-   * @brief Get the length of the buffer result array.
-   * @return The length of the buffer result array.
-   */
-  const int getResultsLength() const override;
-  /**
-   * @brief Get the location of result with respect
-   * to the frame generated by the input device.
-   * @param[in] idx The index of the result.
-   */
-  const dynamic_vino_lib::Result * getLocationResult(int idx) const override;
-  /**
-   * @brief Show the observed detection result either through image window
-     or ROS topic.
-   */
-  const void observeOutput(const std::shared_ptr<Outputs::BaseOutput> & output);
-  /**
-   * @brief Get the name of the Inference instance.
-   * @return The name of the Inference instance.
-   */
-  const std::string getName() const override;
-
-private:
-  std::shared_ptr<Models::ObjectDetectionSSDModel> valid_model_;
-  std::vector<Result> raw_results_;
-  std::vector<Result> results_;
-  int width_ = 0;
-  int height_ = 0;
-  int max_proposal_count_;
-  int object_size_;
-  double show_output_thresh_ = 0;
-
-  int max_batch_size_ = 1;
-  cv::Rect ROI_;
-
-  bool enable_roi_constraint_ = false;
-
-  int getEntryIndex(int side, int lcoords, int lclasses, int location, int entry);
-  double IntersectionOverUnion(const Result &box_1, const Result &box_2);
-  bool matToBlob(const cv::Mat& frame, const cv::Rect&, float scale_factor,
-                 int batch_index, const std::string& input_name);
-};
-}  // namespace dynamic_vino_lib
-#endif  // DYNAMIC_VINO_LIB__INFERENCES__OBJECT_DETECTION_SSD_HPP_
